remove code for creating buckets/siteconfigs
the s3 doesn't let us specify a website config without including things
like the error page and redirects (which we don't know), and creating a
bucket is really a one-time thing... not something a deployment tool
should nessissarly be dealing with. maybe we could add this stuff back
later but that website config bug needs to be fixed first.